Refactor: split page.tsx into modular components (#56)
Extract 8 components from monolithic page files into separate modules:

- LoginForm.tsx: login/register form with brand panel
- AddressSearch.tsx: building lookup component
- ProjectCard.tsx: individual project card
- TemplateGrid.tsx: template picker section
- ProjectList.tsx: project list with nav bar
- SceneEditor.tsx: scene code editor textarea
- BomPanel.tsx: bill of materials panel
- ChatPanel.tsx: AI chat assistant panel

Shared types moved to web/src/types/index.ts.
page.tsx reduced from 875 to ~73 lines.
project/[id]/page.tsx reduced from 1054 to ~576 lines.
No visual or behavioral changes.

Closes #56
