+++ web/src/app/project/[id]/page.tsx
@@ -8,6 +8,10 @@ import { useToast } from "@/components/ToastProvider";
 import { SkeletonProjectEditor } from "@/components/Skeleton";
 import { useTranslation } from "@/components/LocaleProvider";
 import { LanguageSwitcher } from "@/components/LanguageSwitcher";
+import SceneEditor from "@/components/SceneEditor";
+import BomPanel from "@/components/BomPanel";
+import ChatPanel from "@/components/ChatPanel";
+import type { Material, BomItem, Project } from "@/types";
 
 const Viewport3D = dynamic(() => import("@/components/Viewport3D"), {
   ssr: false,
@@ -17,491 +21,6 @@ const Viewport3D = dynamic(() => import("@/components/Viewport3D"), {
     </div>
   ),
 });
-
-interface Material {
-  id: string;
-  name: string;
-  category_name: string;
-  image_url: string | null;
-  pricing: { unit_price: number; unit: string; supplier_name: string; is_primary: boolean }[] | null;
-}
-
-interface BomItem {
-  id?: string;
-  material_id: string;
-  material_name?: string;
-  image_url?: string | null;
-  quantity: number;
-  unit: string;
-  unit_price?: number;
-  total?: number;
-  supplier?: string;
-}
-
-interface Project {
-  id: string;
-  name: string;
-  description: string;
-  scene_js: string | null;
-  display_scale: number;
-  estimated_cost: number;
-  updated_at: string;
-  bom?: BomItem[];
-}
-
-function SceneEditor({
-  sceneJs,
-  onChange,
-}: {
-  sceneJs: string;
-  onChange: (code: string) => void;
-}) {
-  const { t } = useTranslation();
-
-  return (
-    <div style={{ flex: 1, display: "flex", flexDirection: "column" }}>
-      <div style={{ display: "flex", alignItems: "center", gap: 8, marginBottom: 8 }}>
-        <div style={{ width: 8, height: 8, borderRadius: "50%", background: "var(--success)" }} />
-        <span style={{ fontSize: 12, fontWeight: 500, color: "var(--text-muted)", textTransform: "uppercase", letterSpacing: "0.05em" }}>
-          {t('editor.scene')}
-        </span>
-      </div>
-      <textarea
-        value={sceneJs}
-        onChange={(e) => onChange(e.target.value)}
-        spellCheck={false}
-        style={{
-          flex: 1,
-          fontFamily: "var(--font-mono)",
-          fontSize: 13,
-          lineHeight: 1.7,
-          padding: 20,
-          border: "1px solid var(--border)",
-          borderRadius: "var(--radius-md)",
-          resize: "none",
-          background: "var(--bg-tertiary)",
-          color: "#cdd6f4",
-          outline: "none",
-          tabSize: 2,
-        }}
-        onKeyDown={(e) => {
-          if (e.key === "Tab") {
-            e.preventDefault();
-            const target = e.target as HTMLTextAreaElement;
-            const start = target.selectionStart;
-            const end = target.selectionEnd;
-            const val = target.value;
-            onChange(val.substring(0, start) + "  " + val.substring(end));
-            setTimeout(() => {
-              target.selectionStart = target.selectionEnd = start + 2;
-            }, 0);
-          }
-        }}
-      />
-    </div>
-  );
-}
-
-function BomPanel({
-  bom,
-  materials,
-  onAdd,
-  onRemove,
-  onUpdateQty,
-}: {
-  bom: BomItem[];
-  materials: Material[];
-  onAdd: (materialId: string, qty: number) => void;
-  onRemove: (materialId: string) => void;
-  onUpdateQty: (materialId: string, qty: number) => void;
-}) {
-  const [selectedMat, setSelectedMat] = useState("");
-  const [qty, setQty] = useState(1);
-  const { t } = useTranslation();
-
-  const total = bom.reduce((sum, item) => sum + (item.total || 0), 0);
-
-  return (
-    <div
-      style={{
-        width: 360,
-        borderLeft: "1px solid var(--border)",
-        display: "flex",
-        flexDirection: "column",
-        background: "var(--bg-secondary)",
-      }}
-    >
-      <div style={{ padding: "16px 20px", borderBottom: "1px solid var(--border)" }}>
-        <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: 12 }}>
-          <h3 style={{ margin: 0, fontSize: 14, fontWeight: 600 }}>{t('editor.materialList')}</h3>
-          <span className="label-mono" style={{ fontSize: 10, color: "var(--text-muted)" }}>
-            {bom.length} {bom.length === 1 ? 'rivi' : 'rivia'}
-          </span>
-        </div>
-        <div style={{
-          padding: "14px 16px",
-          background: "linear-gradient(135deg, rgba(196,145,92,0.12) 0%, rgba(196,145,92,0.04) 100%)",
-          borderRadius: "var(--radius-md)",
-          border: "1px solid var(--amber-border)",
-        }}>
-          <div className="label-mono" style={{ fontSize: 10, color: "var(--amber)", marginBottom: 6 }}>
-            ARVIOITU KOKONAISHINTA
-          </div>
-          <div style={{ display: "flex", alignItems: "baseline", gap: 6 }}>
-            <span className="heading-display" style={{ fontSize: 24, color: "var(--text-primary)" }}>
-              {total > 0 ? total.toLocaleString('fi-FI', { maximumFractionDigits: 0 }) : '0'}
-            </span>
-            <span style={{ fontSize: 14, color: "var(--text-muted)" }}>&euro;</span>
-            {total > 0 && (
-              <span style={{ marginLeft: "auto", fontSize: 11, color: "var(--text-muted)", fontFamily: "var(--font-mono)" }}>
-                sis. ALV 25.5%
-              </span>
-            )}
-          </div>
-        </div>
-      </div>
-
-      <div style={{ flex: 1, overflowY: "auto", padding: 12 }}>
-        {bom.length === 0 ? (
-          <div style={{ textAlign: "center", padding: "32px 16px" }}>
-            <div style={{ color: "var(--text-muted)", fontSize: 13 }}>
-              {t('editor.noMaterials')}
-            </div>
-          </div>
-        ) : (
-          bom.map((item) => (
-            <div
-              key={item.material_id}
-              style={{
-                padding: "12px 14px",
-                background: "var(--bg-tertiary)",
-                borderRadius: "var(--radius-sm)",
-                marginBottom: 6,
-                fontSize: 13,
-                border: "1px solid var(--border)",
-              }}
-            >
-              <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center", gap: 8 }}>
-                <div style={{ display: "flex", alignItems: "center", gap: 8 }}>
-                  {item.image_url ? (
-                    <img
-                      src={item.image_url}
-                      alt={item.material_name || ""}
-                      style={{ width: 28, height: 28, borderRadius: 4, objectFit: "cover", flexShrink: 0 }}
-                    />
-                  ) : (
-                    <div style={{ width: 28, height: 28, borderRadius: 4, background: "var(--bg-elevated)", flexShrink: 0 }} />
-                  )}
-                  <strong style={{ fontSize: 13, fontWeight: 500 }}>{item.material_name}</strong>
-                </div>
-                <button
-                  onClick={() => onRemove(item.material_id)}
-                  style={{
-                    background: "none",
-                    border: "none",
-                    color: "var(--danger)",
-                    cursor: "pointer",
-                    fontSize: 14,
-                    padding: "0 4px",
-                    opacity: 0.6,
-                  }}
-                  onMouseEnter={(e) => { e.currentTarget.style.opacity = "1"; }}
-                  onMouseLeave={(e) => { e.currentTarget.style.opacity = "0.6"; }}
-                >
-                  x
-                </button>
-              </div>
-              <div style={{ display: "flex", alignItems: "center", gap: 8, marginTop: 8 }}>
-                <input
-                  type="number"
-                  min={0.01}
-                  step={0.1}
-                  value={item.quantity}
-                  onChange={(e) =>
-                    onUpdateQty(item.material_id, parseFloat(e.target.value) || 0)
-                  }
-                  style={{
-                    width: 56,
-                    padding: "4px 6px",
-                    background: "var(--bg-elevated)",
-                    border: "1px solid var(--border)",
-                    borderRadius: 4,
-                    fontSize: 12,
-                    color: "var(--text-primary)",
-                    outline: "none",
-                    fontFamily: "var(--font-mono)",
-                  }}
-                />
-                <span style={{ color: "var(--text-muted)", fontSize: 12 }}>
-                  {item.unit} x {(item.unit_price || 0).toFixed(2)}
-                </span>
-                <span style={{ marginLeft: "auto", fontWeight: 600, color: "var(--success)", fontFamily: "var(--font-mono)", fontSize: 12 }}>
-                  {(item.total || 0).toFixed(2)}
-                </span>
-              </div>
-              {item.supplier && (
-                <div style={{ color: "var(--text-muted)", fontSize: 11, marginTop: 6 }}>
-                  {item.supplier}
-                </div>
-              )}
-            </div>
-          ))
-        )}
-      </div>
-
-      <div
-        style={{
-          padding: 12,
-          borderTop: "1px solid var(--border)",
-          display: "flex",
-          gap: 6,
-          alignItems: "center",
-        }}
-      >
-        <select
-          value={selectedMat}
-          onChange={(e) => setSelectedMat(e.target.value)}
-          style={{
-            flex: 1,
-            padding: "7px 8px",
-            background: "var(--bg-tertiary)",
-            border: "1px solid var(--border)",
-            borderRadius: "var(--radius-sm)",
-            fontSize: 12,
-            color: "var(--text-primary)",
-            outline: "none",
-          }}
-        >
-          <option value="">{t('editor.addMaterial')}</option>
-          {materials
-            .filter((m) => !bom.some((b) => b.material_id === m.id))
-            .map((m) => (
-              <option key={m.id} value={m.id}>
-                {m.name}
-              </option>
-            ))}
-        </select>
-        <input
-          type="number"
-          min={1}
-          value={qty}
-          onChange={(e) => setQty(parseInt(e.target.value) || 1)}
-          style={{
-            width: 48,
-            padding: "7px 6px",
-            background: "var(--bg-tertiary)",
-            border: "1px solid var(--border)",
-            borderRadius: "var(--radius-sm)",
-            fontSize: 12,
-            color: "var(--text-primary)",
-            outline: "none",
-            fontFamily: "var(--font-mono)",
-          }}
-        />
-        <button
-          className={`btn ${selectedMat ? "btn-primary" : ""}`}
-          onClick={() => {
-            if (selectedMat) {
-              onAdd(selectedMat, qty);
-              setSelectedMat("");
-              setQty(1);
-            }
-          }}
-          disabled={!selectedMat}
-          style={{
-            padding: "7px 14px",
-            fontSize: 12,
-            opacity: selectedMat ? 1 : 0.4,
-          }}
-        >
-          {t('editor.add')}
-        </button>
-      </div>
-    </div>
-  );
-}
-
-interface ChatMessage {
-  role: "user" | "assistant";
-  content: string;
-}
-
-function ChatPanel({
-  sceneJs,
-  onApplyCode,
-}: {
-  sceneJs: string;
-  onApplyCode: (code: string) => void;
-}) {
-  const [messages, setMessages] = useState<ChatMessage[]>([]);
-  const [input, setInput] = useState("");
-  const [loading, setLoading] = useState(false);
-  const { toast } = useToast();
-  const { t } = useTranslation();
-
-  async function send() {
-    if (!input.trim() || loading) return;
-    const userMsg: ChatMessage = { role: "user", content: input };
-    const newMessages = [...messages, userMsg];
-    setMessages(newMessages);
-    setInput("");
-    setLoading(true);
-
-    try {
-      const reply = await api.chat(newMessages, sceneJs);
-      setMessages([...newMessages, reply]);
-    } catch (err) {
-      toast(err instanceof Error ? err.message : t('toast.aiError'), "error");
-      setMessages([
-        ...newMessages,
-        { role: "assistant", content: t('editor.chatError') },
-      ]);
-    }
-    setLoading(false);
-  }
-
-  function extractCode(content: string): string | null {
-    const match = content.match(/```(?:javascript|js)?\n([\s\S]*?)```/);
-    return match ? match[1].trim() : null;
-  }
-
-  return (
-    <div
-      className="animate-slide"
-      style={{
-        display: "flex",
-        flexDirection: "column",
-        height: "100%",
-        background: "var(--bg-secondary)",
-      }}
-    >
-      <div
-        style={{
-          padding: "14px 16px",
-          borderBottom: "1px solid var(--border)",
-          fontSize: 13,
-          fontWeight: 600,
-          display: "flex",
-          alignItems: "center",
-          gap: 8,
-        }}
-      >
-        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="var(--accent)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-          <path d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z" />
-        </svg>
-        {t('editor.aiAssistant')}
-      </div>
-      <div
-        style={{
-          flex: 1,
-          overflowY: "auto",
-          padding: 12,
-          display: "flex",
-          flexDirection: "column",
-          gap: 8,
-        }}
-      >
-        {messages.length === 0 && (
-          <div style={{ color: "var(--text-muted)", fontSize: 13, padding: "24px 12px", textAlign: "center", lineHeight: 2 }}>
-            {t('editor.describePrompt')}
-            <br />
-            <span style={{ color: "#c4915c", opacity: 0.7 }}>
-              &ldquo;{t('editor.exampleRoof')}&rdquo;
-            </span>
-            <br />
-            <span style={{ color: "#c4915c", opacity: 0.7 }}>
-              &ldquo;{t('editor.exampleWindow')}&rdquo;
-            </span>
-          </div>
-        )}
-        {messages.map((msg, i) => {
-          const code = msg.role === "assistant" ? extractCode(msg.content) : null;
-          const textContent = msg.content
-            .replace(/```(?:javascript|js)?\n[\s\S]*?```/g, "[code block]")
-            .trim();
-
-          return (
-            <div
-              key={i}
-              className="animate-in"
-              style={{
-                alignSelf: msg.role === "user" ? "flex-end" : "flex-start",
-                maxWidth: "88%",
-              }}
-            >
-              <div
-                style={{
-                  padding: "8px 12px",
-                  borderRadius: msg.role === "user" ? "12px 12px 4px 12px" : "12px 12px 12px 4px",
-                  fontSize: 13,
-                  lineHeight: 1.5,
-                  background: msg.role === "user" ? "var(--accent)" : "var(--bg-elevated)",
-                  color: msg.role === "user" ? "#fff" : "var(--text-primary)",
-                  whiteSpace: "pre-wrap",
-                }}
-              >
-                {textContent}
-              </div>
-              {code && (
-                <button
-                  className="btn"
-                  onClick={() => onApplyCode(code)}
-                  style={{
-                    marginTop: 6,
-                    padding: "4px 10px",
-                    background: "var(--success-muted)",
-                    color: "var(--success)",
-                    fontSize: 11,
-                    fontWeight: 600,
-                  }}
-                >
-                  {t('editor.applyToScene')}
-                </button>
-              )}
-            </div>
-          );
-        })}
-        {loading && (
-          <div style={{ color: "var(--accent)", fontSize: 13, padding: 8 }}>
-            <span style={{ animation: "pulse 1.5s infinite" }}>{t('editor.thinking')}</span>
-          </div>
-        )}
-      </div>
-      <div
-        style={{
-          padding: 12,
-          borderTop: "1px solid var(--border)",
-          display: "flex",
-          gap: 6,
-        }}
-      >
-        <input
-          className="input"
-          value={input}
-          onChange={(e) => setInput(e.target.value)}
-          onKeyDown={(e) => e.key === "Enter" && !e.shiftKey && send()}
-          placeholder={t('editor.describeChange')}
-          style={{ flex: 1, padding: "8px 12px", fontSize: 13 }}
-        />
-        <button
-          className="btn btn-primary"
-          onClick={send}
-          disabled={loading || !input.trim()}
-          style={{
-            padding: "8px 14px",
-            fontSize: 13,
-            opacity: loading || !input.trim() ? 0.4 : 1,
-          }}
-        >
-          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-            <line x1="22" y1="2" x2="11" y2="13" />
-            <polygon points="22 2 15 22 11 13 2 9 22 2" />
-          </svg>
-        </button>
-      </div>
-    </div>
-  );
-}
 
 const DEFAULT_SCENE = `// Helscoop Scene Script
 // Available: box(w,h,d), cylinder(r,h), sphere(r)
@@ -543,33 +62,24 @@ export default function ProjectPage() {
   const [projectName, setProjectName] = useState("");
   const [projectDesc, setProjectDesc] = useState("");
   const [showChat, setShowChat] = useState(false);
-  const [showExport, setShowExport] = useState(false);
-  const exportDropdownRef = useRef<HTMLDivElement>(null);
   const [showCode, setShowCode] = useState(false);
   const [wireframe, setWireframe] = useState(false);
   const [objectCount, setObjectCount] = useState(0);
   const [sceneError, setSceneError] = useState<string | null>(null);
   const viewportRef = useRef<HTMLDivElement>(null);
 
-  // Undo/redo history for scene script
   const historyRef = useRef<string[]>([]);
   const historyIndexRef = useRef<number>(-1);
 
-  // Debounce timer for auto-save
   const autoSaveTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  // Track whether BOM changed since last save
   const bomChangedRef = useRef(false);
-  // Track whether initial load is done (to avoid auto-saving on mount)
   const initialLoadDoneRef = useRef(false);
 
-  // Push a scene script entry to the history stack
   const pushHistory = useCallback((code: string) => {
     const history = historyRef.current;
     const idx = historyIndexRef.current;
-    // Remove any forward history when a new change is made
     const newHistory = history.slice(0, idx + 1);
     newHistory.push(code);
-    // Enforce limit
     if (newHistory.length > HISTORY_LIMIT) {
       newHistory.shift();
     }
@@ -606,12 +116,10 @@ export default function ProjectPage() {
         setProjectDesc(proj.description || "");
         const initialScene = proj.scene_js || DEFAULT_SCENE;
         setSceneJs(initialScene);
-        // Initialize history with the loaded scene
         historyRef.current = [initialScene];
         historyIndexRef.current = 0;
         setMaterials(mats);
         if (proj.bom) setBom(proj.bom);
-        // Mark initial load as done after a tick so auto-save doesn't fire for initial state
         setTimeout(() => {
           initialLoadDoneRef.current = true;
         }, 0);
@@ -627,7 +135,6 @@ export default function ProjectPage() {
       });
   }, [projectId, router, toast, t]);
 
-  // Save function (used by both auto-save and manual save)
   const save = useCallback(async () => {
     setSaveStatus("saving");
     try {
@@ -661,7 +168,6 @@ export default function ProjectPage() {
     }
   }, [projectId, projectName, projectDesc, sceneJs, bom, toast, t]);
 
-  // Schedule auto-save (debounced 2 seconds)
   const scheduleAutoSave = useCallback(() => {
     if (!initialLoadDoneRef.current) return;
     setSaveStatus("unsaved");
@@ -673,7 +179,6 @@ export default function ProjectPage() {
     }, 2000);
   }, [save]);
 
-  // Auto-save when project data changes
   useEffect(() => {
     if (!initialLoadDoneRef.current) return;
     scheduleAutoSave();
@@ -684,7 +189,6 @@ export default function ProjectPage() {
     };
   }, [sceneJs, projectName, projectDesc, bom, scheduleAutoSave]);
 
-  // beforeunload warning for unsaved changes
   useEffect(() => {
     const handler = (e: BeforeUnloadEvent) => {
       if (saveStatus === "unsaved" || saveStatus === "saving") {
@@ -695,19 +199,6 @@ export default function ProjectPage() {
     return () => window.removeEventListener("beforeunload", handler);
   }, [saveStatus]);
 
-  // Close export dropdown on click outside
-  useEffect(() => {
-    if (!showExport) return;
-    const handler = (e: MouseEvent) => {
-      if (exportDropdownRef.current && !exportDropdownRef.current.contains(e.target as Node)) {
-        setShowExport(false);
-      }
-    };
-    document.addEventListener("mousedown", handler);
-    return () => document.removeEventListener("mousedown", handler);
-  }, [showExport]);
-
-  // Keyboard shortcuts for undo/redo
   useEffect(() => {
     const handler = (e: KeyboardEvent) => {
       const isMod = e.metaKey || e.ctrlKey;
@@ -723,7 +214,6 @@ export default function ProjectPage() {
     return () => window.removeEventListener("keydown", handler);
   }, [undo, redo]);
 
-  // Handle scene changes: push to history and update state
   const handleSceneChange = useCallback(
     (code: string) => {
       setSceneJs(code);
@@ -732,7 +222,6 @@ export default function ProjectPage() {
     [pushHistory]
   );
 
-  // Handle AI chat applying code: push to history as single unit
   const handleApplyCode = useCallback(
     (code: string) => {
       setSceneJs(code);
@@ -849,7 +338,6 @@ export default function ProjectPage() {
             width: 180,
           }}
         />
-        {/* Undo/Redo buttons */}
         <div style={{ display: "flex", gap: 2 }}>
           <button
             className="btn btn-ghost"
@@ -903,71 +391,23 @@ export default function ProjectPage() {
         <button className="btn" onClick={save} style={{ padding: "6px 16px", background: "linear-gradient(135deg, #c4915c 0%, #a67745 100%)", color: "#fff", border: "none" }}>
           {t('editor.save')}
         </button>
-        <div ref={exportDropdownRef} style={{ position: 'relative' }}>
-          <button
-            className="btn btn-ghost"
-            onClick={() => setShowExport(!showExport)}
-            style={{ padding: "6px 12px", display: "flex", alignItems: "center", gap: 4 }}
-          >
-            {t('editor.export')} <span style={{ fontSize: 10 }}>{"\u25BE"}</span>
-          </button>
-          {showExport && (
-            <div style={{
-              position: 'absolute', top: '100%', right: 0, marginTop: 4,
-              background: 'var(--bg-elevated)', border: '1px solid var(--border)',
-              borderRadius: 'var(--radius-sm)', overflow: 'hidden', zIndex: 10,
-              minWidth: 160, boxShadow: 'var(--shadow-md)',
-            }}>
-              <button
-                onClick={async () => {
-                  setShowExport(false);
-                  try {
-                    await api.exportBOMCsv(projectId, projectName);
-                    toast(t('toast.bomExported'), "success");
-                  } catch (err) {
-                    toast(err instanceof Error ? err.message : t('toast.bomExportFailed'), "error");
-                  }
-                }}
-                style={{
-                  display: 'block', width: '100%', textAlign: 'left',
-                  padding: '10px 16px', fontSize: 13, border: 'none',
-                  background: 'transparent', color: 'var(--text-primary)', cursor: 'pointer',
-                }}
-                onMouseEnter={(e) => { e.currentTarget.style.background = 'var(--bg-hover)'; }}
-                onMouseLeave={(e) => { e.currentTarget.style.background = 'transparent'; }}
-              >
-                CSV (Excel)
-              </button>
-              <button
-                onClick={async () => {
-                  setShowExport(false);
-                  try {
-                    const res = await api.exportBOM(projectId);
-                    const blob = new Blob([JSON.stringify(res, null, 2)], { type: "application/json" });
-                    const url = URL.createObjectURL(blob);
-                    const a = document.createElement("a");
-                    a.href = url;
-                    a.download = `bom_${projectId}.json`;
-                    a.click();
-                    URL.revokeObjectURL(url);
-                    toast(t('toast.bomExported'), "success");
-                  } catch (err) {
-                    toast(err instanceof Error ? err.message : t('toast.bomExportFailed'), "error");
-                  }
-                }}
-                style={{
-                  display: 'block', width: '100%', textAlign: 'left',
-                  padding: '10px 16px', fontSize: 13, border: 'none',
-                  background: 'transparent', color: 'var(--text-primary)', cursor: 'pointer',
-                }}
-                onMouseEnter={(e) => { e.currentTarget.style.background = 'var(--bg-hover)'; }}
-                onMouseLeave={(e) => { e.currentTarget.style.background = 'transparent'; }}
-              >
-                JSON
-              </button>
-            </div>
-          )}
-        </div>
+        <button className="btn btn-ghost" onClick={async () => {
+          try {
+            const res = await api.exportBOM(projectId);
+            const blob = new Blob([JSON.stringify(res, null, 2)], { type: "application/json" });
+            const url = URL.createObjectURL(blob);
+            const a = document.createElement("a");
+            a.href = url;
+            a.download = `bom_${projectId}.json`;
+            a.click();
+            URL.revokeObjectURL(url);
+            toast(t('toast.bomExported'), "success");
+          } catch (err) {
+            toast(err instanceof Error ? err.message : t('toast.bomExportFailed'), "error");
+          }
+        }}>
+          {t('editor.export')}
+        </button>
         <button
           className="btn"
           onClick={() => setShowChat(!showChat)}
@@ -987,7 +427,6 @@ export default function ProjectPage() {
 
       {/* Main content */}
       <div style={{ flex: 1, display: "flex", overflow: "hidden" }}>
-        {/* Left: Viewport + Code */}
         <div
           style={{
             flex: 1,
