+++ web/src/types/index.ts
@@ -1,0 +1,63 @@
+export interface Project {
+  id: string;
+  name: string;
+  description: string;
+  estimated_cost: number;
+  updated_at: string;
+  scene_js?: string | null;
+  display_scale?: number;
+  bom?: BomItem[];
+}
+
+export interface BuildingResult {
+  address: string;
+  coordinates: { lat: number; lon: number };
+  building_info: {
+    type: string;
+    year_built: number;
+    material: string;
+    floors: number;
+    area_m2: number;
+    heating: string;
+    roof_type?: string;
+    roof_material?: string;
+    units?: number;
+  };
+  scene_js: string;
+  bom_suggestion: { material_id: string; quantity: number; unit: string }[];
+}
+
+export interface Template {
+  id: string;
+  name: string;
+  description: string;
+  icon: string;
+  estimated_cost: number;
+  scene_js: string;
+  bom: { material_id: string; quantity: number; unit: string }[];
+}
+
+export interface Material {
+  id: string;
+  name: string;
+  category_name: string;
+  image_url: string | null;
+  pricing: { unit_price: number; unit: string; supplier_name: string; is_primary: boolean }[] | null;
+}
+
+export interface BomItem {
+  id?: string;
+  material_id: string;
+  material_name?: string;
+  image_url?: string | null;
+  quantity: number;
+  unit: string;
+  unit_price?: number;
+  total?: number;
+  supplier?: string;
+}
+
+export interface ChatMessage {
+  role: "user" | "assistant";
+  content: string;
+}
